src: move credential code to sub-package
