+++ cmd/build.go
@@ -12,6 +12,7 @@ import (
 	fn "knative.dev/kn-plugin-func"
 	"knative.dev/kn-plugin-func/buildpacks"
 	"knative.dev/kn-plugin-func/docker"
+	"knative.dev/kn-plugin-func/docker/creds"
 	"knative.dev/kn-plugin-func/progress"
 )
 
@@ -27,9 +28,9 @@ func newBuildClient(cfg buildConfig) (*fn.Client, error) {
 
 	pusherOption := fn.WithPusher(nil)
 	if cfg.Push {
-		credentialsProvider := docker.NewCredentialsProvider(
-			docker.WithPromptForCredentials(newPromptForCredentials()),
-			docker.WithPromptForCredentialStore(newPromptForCredentialStore()),
+		credentialsProvider := creds.NewCredentialsProvider(
+			creds.WithPromptForCredentials(newPromptForCredentials()),
+			creds.WithPromptForCredentialStore(newPromptForCredentialStore()),
 		)
 		pusher, err := docker.NewPusher(
 			docker.WithCredentialsProvider(credentialsProvider),
